refactor(email): 重构邮箱验证功能
- 新增 EmailRegisterController 和 EmailForgotController 分别处理注册和重置密码的邮箱验证- 更新前端代码，使用新的 API 接口
- 重构后端逻辑，提高代码复用性和可维护性
- 优化国际化支持，增加新的错误消息

+++ src/main/java/com/jiang/mall/service/IUserService.java
@@ -33,7 +33,7 @@ public interface IUserService extends IService<User> {
 
     User login(String username, String password);
 
-    Boolean modifyPassword(Integer userId, String oldPassword, String newPassword);
+    Boolean modifyPassword(Long userId, String oldPassword, String newPassword);
 
     User getUserInfo(Long userId);
